backend/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
